Xtro.MDX:
* Size bug is fixed in Object.GetPrivateData method for x64 bit.

+++ Libraries/Xtro.MDX/BlendDescription.h
@@ -25,7 +25,7 @@ public:
 	virtual int GetHashCode() override
 	{
 		return
-			AlphaToCoverageEnable ^
+			(int)AlphaToCoverageEnable ^
 			BlendEnable.GetHashCode() ^
 			(int)SourceBlend ^
 			(int)DestinationBlend ^
